build: Use tsc to build the application, instead of exit-code on error
This commit changes the npm build script to use 'tsc' instead of 'exit 0' when compiling the application. Additionally, it adds validators to the User model for date of birth, adds 'ENTER_VALID_DOB' message to the user resource strings file and includes 'EMPTY_DOB' to the authentication resource strings file.